requirements.txt
flask
flask_restful
flask_jwt_extended
flask_sqlalchemy